refactor: extract media and renderer helpers
